Starting out with small refactorings
