+++ src/main/python/crassus/deployer.py
@@ -3,6 +3,7 @@ import logging
 
 import boto3
 from botocore.exceptions import ClientError
+from crassus.result_message import ResultMessage
 
 logger = logging.getLogger('crassus-deployer')
 logger.setLevel(logging.DEBUG)
@@ -12,9 +13,6 @@ logger.addHandler(consoleLogger)
 NOTIFICATION_SUBJECT = 'Crassus deployer notification'
 MESSAGE_STACK_NOT_FOUND = 'Stack not found {stack_name}: {message}'
 MESSAGE_UPDATE_PROBLEM = 'Problem while updating stack {stack_name}: {message}'
-
-STATUS_SUCCESS = 'success'
-STATUS_FAILURE = 'failure'
 
 
 class Crassus(object):
@@ -94,7 +92,7 @@ class Crassus(object):
         except ClientError as error:
             logger.error(MESSAGE_STACK_NOT_FOUND.format(
                 stack_name=self.stack_name, message=error.message))
-            self.notify(STATUS_FAILURE, error.message)
+            self.notify(ResultMessage.STATUS_FAILURE, error.message)
 
     def update(self):
         merged = self.stack_update_parameters.merge(self.stack.parameters)
@@ -105,14 +103,14 @@ class Crassus(object):
                 Parameters=merged,
                 Capabilities=['CAPABILITY_IAM'],
                 NotificationARNs=None)
-                #NotificationARNs=self.output_topics)
+                # NotificationARNs=self.output_topics)
             message = 'Cloudformation was triggered successfully.'
             logger.debug(message)
-            self.notify(STATUS_SUCCESS, message)
+            self.notify(ResultMessage.STATUS_SUCCESS, message)
         except ClientError as error:
             logger.error(MESSAGE_UPDATE_PROBLEM.format(
                 stack_name=self.stack_name, message=error.message))
-            self.notify(STATUS_FAILURE, error.message)
+            self.notify(ResultMessage.STATUS_FAILURE, error.message)
 
     def deploy(self):
         self.load()
@@ -147,27 +145,3 @@ class StackUpdateParameter(dict):
                     merged_stack_parameters.append(stack_parameter)
 
         return merged_stack_parameters
-
-
-class ResultMessage(dict):
-
-    """
-    A message that crassus returns for events such as fail or success
-    events for stack deployments/updates. These messages will be
-    transmitted as JSON encoded strings.
-
-    It is initialized with the following parameters:
-    - status: STATUS_FAILURE or STATUS_SUCCESS
-    - stack_name: the stack name the notification message stands for
-    - version: a version identifier for the message. Defaults as '1.0'
-    - message: the textual message for the notification.
-    """
-
-    version = '1.0'
-
-    def __init__(self, status, message, stack_name):
-        self['version'] = self.version
-        self['stackName'] = stack_name
-        self['status'] = status
-        self['message'] = message
-
